bench_timers: Split test implementations into separate files
